refactor: ajusta diretórios
