+++ src/app/(aplicacao)/resultado/votacao/page.tsx
